fix segfault when connect() is called without cb function
It is unclear what the original intent in the case of a missing callback
function was. Should the method still try to connect to the cluster?
Should it abort with an error?

Instead of causing a segmentation fault we'll just throw an error for
now. Ideally, the client object should implement the EventEmitter
interface with events for `connected` and `error`. Then the cb function
parameter on the connect() method could be optional; if provided, the
method would just setup the cb as event handler for the client's events.