add file I/O based test case; capture memory usage
Adds a third test scenario to the query/scan performance tests. In
addtion to testing with a noop data event handler to perform query/scan as
fast as possible and a event handler with busy loop blocking processing
on the event loop, the 3rd test scenario performs some asynchronous file
I/O in the data event handler.

process.memoryUsage() is used to capture RSS and heap usage during the
tests.